refactor: move auth to lib
